requirements.txt
accelerate==0.33.0
datasets==2.21.0
evaluate==0.4.2
flash-attn==2.7.2.post1
huggingface-hub==0.24.6
hydra-core==1.3.2
intervaltree==3.1.0
jiwer==2.5.2
kaldiio==2.18.0
lhotse==1.28.0
librosa==0.10.2.post1
matplotlib==3.9.2
matplotlib-inline==0.1.7
meeteval==0.3.0
more-itertools==10.2.0
ninja==1.11.1.2
numpy==1.26.4
pandas==2.2.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
PyYAML==6.0.2
safetensors==0.4.4
scipy==1.14.1
soundfile==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
tqdm==4.66.5
transformers==4.41.2
wandb==0.19.0